Garden App debugged

+++ week-04/day-1/src/GardenApp/Garden.java
@@ -34,7 +34,7 @@ public class Garden {
     public void waterGarden(int waterAmount) {
 
         double waterShare;
-        waterShare = waterAmount / plants.size();
+        waterShare = (double)(waterAmount) / (double)(plants.size());
 
         for (Plant item : plants) {
